pyproject.toml
[tool.poetry]
name = "water-dp"
version = "0.1.0"
description = "Water Data Platform backend (FastAPI + SQLAlchemy) with GeoServer and time series support."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
  { include = "app" },
  { include = "scripts" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
alembic = "^1.13.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.0"
structlog = "^24.0.0"
prometheus-client = "^0.20.0"
pandas = "^2.2.0"
numpy = "^1.26.0"
pytz = "^2024.1"
requests = "^2.32.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.28.0"
python-multipart = "^0.0.9"
python-dotenv = "^1.0.0"
geoalchemy2 = "^0.14.0"
shapely = "^2.1.2"
asyncpg = "^0.31.0"
celery = "^5.6.2"
redis = "^7.1.0"
python-keycloak = "^4.0.0"
paho-mqtt = "^1.6.1"
minio = "^7.2.0"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.0"
mypy = "^1.9.0"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.0"
ruff = "^0.4.0"

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []
extend-ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app", "scripts"]


[tool.poetry.scripts]
start-server = "start_server:start_server"
run-migrations = "scripts.run_migrations:main"
fix = "scripts.fix_code:main"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"


